refactor: Move the useStepperDialog hook to its Provider file
