--- src/components/Hooks/useStepperDialog.jsx
@@ -1,12 +1,0 @@
-import { useContext } from 'react'
-import StepperDialogContext from 'src/components/Contexts/StepperDialogProvider'
-
-export const useStepperDialog = () => {
-  const stepperDialogContext = useContext(StepperDialogContext)
-  if (!stepperDialogContext) {
-    throw new Error(
-      'useStepperDialog must be used within a StepperDialogProvider'
-    )
-  }
-  return stepperDialogContext
-}
