+++ src/components/ModelSteps/ContactDialog.jsx
@@ -2,7 +2,7 @@ import PropTypes from 'prop-types'
 import React, { useEffect, useState } from 'react'
 import CompositeHeader from 'src/components/CompositeHeader/CompositeHeader'
 import { useFormData } from 'src/components/Contexts/FormDataProvider'
-import { useStepperDialog } from 'src/components/Hooks/useStepperDialog'
+import { useStepperDialog } from 'src/components/Contexts/StepperDialogProvider'
 import ContactList from 'src/components/ModelSteps/ContactList'
 import SubmitButton from 'src/components/ModelSteps/widgets/SubmitButton'
 import StepperDialogTitle from 'src/components/StepperDialog/StepperDialogTitle'
