+++ src/components/ModelSteps/ContactDialog.spec.jsx
@@ -5,7 +5,7 @@ import React from 'react'
 import { FormDataProvider } from 'src/components/Contexts/FormDataProvider'
 import { useFormData } from 'src/components/Contexts/FormDataProvider'
 import { StepperDialogProvider } from 'src/components/Contexts/StepperDialogProvider'
-import { useStepperDialog } from 'src/components/Hooks/useStepperDialog'
+import { useStepperDialog } from 'src/components/Contexts/StepperDialogProvider'
 import ContactDialog from 'src/components/ModelSteps/ContactDialog'
 import { fetchCurrentUser } from 'src/helpers/fetchCurrentUser'
 import AppLike from 'test/components/AppLike'
