+++ src/components/ModelSteps/InformationDialog.jsx
@@ -7,7 +7,7 @@ import IlluGenericInputDate from 'src/assets/icons/IlluGenericInputDate.svg'
 import IlluGenericInputText from 'src/assets/icons/IlluGenericInputText.svg'
 import CompositeHeader from 'src/components/CompositeHeader/CompositeHeader'
 import { useFormData } from 'src/components/Contexts/FormDataProvider'
-import { useStepperDialog } from 'src/components/Hooks/useStepperDialog'
+import { useStepperDialog } from 'src/components/Contexts/StepperDialogProvider'
 import { normalizeFormdataMetadata } from 'src/components/ModelSteps/helpers'
 import SubmitButton from 'src/components/ModelSteps/widgets/SubmitButton'
 import StepperDialogTitle from 'src/components/StepperDialog/StepperDialogTitle'
