+++ src/components/ModelSteps/NoteDialog.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { useParams } from 'react-router-dom'
 import { useFormData } from 'src/components/Contexts/FormDataProvider'
-import { useStepperDialog } from 'src/components/Hooks/useStepperDialog'
+import { useStepperDialog } from 'src/components/Contexts/StepperDialogProvider'
 import StepperDialogTitle from 'src/components/StepperDialog/StepperDialogTitle'
 import { FILES_DOCTYPE } from 'src/constants'
 import { addContactReferenceToFile } from 'src/helpers/createPdfAndSave'
