+++ src/components/ModelSteps/Scan/ScanActions/ScanDesktopActions.jsx
@@ -2,7 +2,7 @@ import PropTypes from 'prop-types'
 import React, { createRef } from 'react'
 import { usePapersCreated } from 'src/components/Contexts/PapersCreatedProvider'
 import { usePaywall } from 'src/components/Contexts/PaywallProvider'
-import { useStepperDialog } from 'src/components/Hooks/useStepperDialog'
+import { useStepperDialog } from 'src/components/Contexts/StepperDialogProvider'
 import ScanDesktopActionsAlert from 'src/components/ModelSteps/Scan/ScanActions/ScanDesktopActionsAlert'
 import { KEYS, SETTINGS_DOCTYPE } from 'src/constants'
 import { getAppSettings } from 'src/helpers/queries'
