+++ src/components/ModelSteps/ScanResult/ScanResultDialog.jsx
@@ -2,7 +2,7 @@ import cx from 'classnames'
 import React, { useRef, useState } from 'react'
 import { useSearchParams } from 'react-router-dom'
 import CompositeHeaderImage from 'src/components/CompositeHeader/CompositeHeaderImage'
-import { useStepperDialog } from 'src/components/Hooks/useStepperDialog'
+import { useStepperDialog } from 'src/components/Contexts/StepperDialogProvider'
 import OcrProcessingDialog from 'src/components/ModelSteps/ScanResult/OcrProcessingDialog'
 import ScanResultCard from 'src/components/ModelSteps/ScanResult/ScanResultCard'
 import ScanResultTitle from 'src/components/ModelSteps/ScanResult/ScanResultTitle'
