+++ src/components/ModelSteps/ScanResult/ScanResultDialog.spec.jsx
@@ -3,7 +3,7 @@ import React from 'react'
 import { FormDataProvider } from 'src/components/Contexts/FormDataProvider'
 import { useFormData } from 'src/components/Contexts/FormDataProvider'
 import { StepperDialogProvider } from 'src/components/Contexts/StepperDialogProvider'
-import { useStepperDialog } from 'src/components/Hooks/useStepperDialog'
+import { useStepperDialog } from 'src/components/Contexts/StepperDialogProvider'
 import ScanResultDialog from 'src/components/ModelSteps/ScanResult/ScanResultDialog'
 import { getAttributesFromOcr } from 'src/components/ModelSteps/helpers'
 import { FLAGSHIP_SCAN_TEMP_FILENAME } from 'src/constants'
